backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Google Gemini
google-generativeai==0.3.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis for session management
redis==5.0.1

# Utilities
python-dateutil==2.8.2
